panel: Add paypal support
This is also a conceptual change, becouse the commit
removes ability to list transfers after uploading and then
marking the owner. We no longer need it, but it might be
introduced in the future, but definetely in a different
form.

Moreover the nasty async.series was removed and the
operation on over thousand rows took a page to load almost
immediately, becouse we no longer wait for processing
transactions and writing the relevant data into database.
All calls are asynchronous now.